+++ lib/src/momentum_base.dart
@@ -2,14 +2,22 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+
+import '../momentum.dart';
 import 'in_memory_storage.dart';
 import 'momentum_error.dart';
 import 'momentum_event.dart';
 import 'momentum_router.dart';
-
 import 'momentum_types.dart';
 
 Type _getType<T>() => T;
+
+MomentumError _invalidService = const MomentumError(
+  'You are not allowed to grab '
+  '"InjectService" directly. You might have done one of these:\n'
+  '1. Momentum.service<InjectService>(...)\n'
+  '2. getService<InjectService>(...)',
+);
 
 /// Used internally.
 /// Simplify trycatch blocks.
@@ -52,6 +60,33 @@ class _MomentumListener<M> {
   _MomentumListener({@required this.state, @required this.invoke});
 }
 
+/// A mixin for [MomentumController] that adds the capability to
+/// handle route changes from momentum's built-in routing system.
+mixin RouterMixin on _ControllerBase {
+  /// Get the current route parameters specified using
+  /// the `params` parameter in `Router.goto(...)` method.
+  ///
+  /// ### Example:
+  /// ```dart
+  /// // setting the route params.
+  /// Router.goto(context, DashboardPage, params: DashboardParams(...));
+  ///
+  /// // accessing the route params inside widgets.
+  /// var params = Router.getParam<DashboardParams>(context);
+  ///
+  /// // accessing the route params inside controllers.
+  /// var params = getParam<DashboardParams>();
+  /// ```
+  T getParam<T extends RouterParam>() {
+    var result = Router.getParam<T>(_mRootContext);
+    return result;
+  }
+
+  /// A callback whenever [Router.goto] or [Router.pop] is called.
+  /// The [RouterParam] is also provided.
+  void onRouteChanged(RouterParam param) {}
+}
+
 /// The class which holds the state of your app.
 /// This is tied with [MomentumController].
 @immutable
@@ -89,11 +124,13 @@ abstract class MomentumModel<Controller extends MomentumController> {
   Map<String, dynamic> toJson() => null;
 }
 
+mixin _ControllerBase {
+  BuildContext _mRootContext;
+}
+
 /// The class which holds the logic for your app.
 /// This is tied with [MomentumModel].
-abstract class MomentumController<M> {
-  BuildContext _mRootContext;
-
+abstract class MomentumController<M> with _ControllerBase {
   /// Dependency injection method for getting other controllers.
   /// Useful for accessing other controllers' function
   /// and model properties without dragging the widget context around.
@@ -118,9 +155,9 @@ abstract class MomentumController<M> {
   /// A method for getting a service marked with
   /// [MomentumService] that are injected into
   /// [Momentum] root widget.
-  T getService<T extends MomentumService>() {
+  T getService<T extends MomentumService>({dynamic alias}) {
     try {
-      var result = Momentum.getService<T>(_mRootContext);
+      var result = Momentum.service<T>(_mRootContext, alias: alias);
       return result;
     } on dynamic catch (e) {
       if (_momentumLogging) {
@@ -681,7 +718,32 @@ abstract class MomentumController<M> {
 /// to inject them into the `services` parameter
 /// of [Momentum] root widget and use them
 /// down the tree.
-abstract class MomentumService {}
+abstract class MomentumService {
+  BuildContext _context;
+
+  /// A method for getting a service marked with
+  /// [MomentumService] that are injected into
+  /// [Momentum] root widget.
+  T getService<T extends MomentumService>({dynamic alias}) {
+    var momentum = Momentum._getMomentumInstance(_context);
+    return momentum._getService<T>(alias: alias);
+  }
+}
+
+/// Use this class to inject multiple services of the
+/// same type with different configurations using alias.
+class InjectService<S extends MomentumService> extends MomentumService {
+  final dynamic _alias;
+  final S _service;
+
+  /// Use this class to inject multiple services of the
+  /// same type with different configurations using alias.
+  InjectService(
+    dynamic alias,
+    S service,
+  )   : _alias = alias,
+        _service = service;
+}
 
 /// A [State] class with additional properties.
 /// Also allows you to add listeners for controllers.
@@ -963,13 +1025,14 @@ class _MomentumRoot extends StatefulWidget {
 }
 
 class _MomentumRootState extends State<_MomentumRoot> {
+  final _momentumEvent = MomentumEvent();
   bool _mErrorFound = false;
 
   Future<bool> _init() async {
+    await _initServices(widget.services);
     await _initControllerModel(widget.controllers);
     _bootstrapControllers(widget.controllers);
     await _bootstrapControllersAsync(widget.controllers);
-    await _initServices(widget.services);
     return true;
   }
 
@@ -994,14 +1057,24 @@ class _MomentumRootState extends State<_MomentumRoot> {
     var momentum = Momentum._getMomentumInstance(context);
     for (var service in services) {
       if (service != null) {
+        service._context = context;
         if (service is Router) {
+          _momentumEvent.add<RouterSignal>().listen((event) {
+            for (var controller in widget.controllers) {
+              if (controller is RouterMixin) {
+                (controller as RouterMixin).onRouteChanged(event.param);
+              }
+            }
+          });
           service.setFunctions(
             context,
             momentum._persistSave,
             momentum._persistGet,
             momentum._syncPersistSave,
             momentum._syncPersistGet,
+            _momentumEvent,
           );
+
           if (widget.testMode) {
             service.initSync(testMode: widget.testMode);
           } else {
@@ -1044,36 +1117,40 @@ class _MomentumRootState extends State<_MomentumRoot> {
     var error = Momentum._getMomentumInstance(context)._validateControllers(
       widget.controllers,
     );
+    error ??= Momentum._getMomentumInstance(context)._validateInjectService(
+      widget.services,
+    );
     if (!_mErrorFound && error != null) {
       _mErrorFound = true;
       throw MomentumError(error);
-    }
-    return FutureBuilder<bool>(
-      future: _init(),
-      initialData: false,
-      builder: (context, snapshot) {
-        if (snapshot.data) {
-          return widget.child;
-        }
-        return widget.appLoader ??
-            MaterialApp(
-              theme: ThemeData(
-                primarySwatch: Colors.blue,
-              ),
-              debugShowCheckedModeBanner: false,
-              home: Scaffold(
-                backgroundColor: Colors.white,
-                body: Center(
-                  child: SizedBox(
-                    height: 36,
-                    width: 36,
-                    child: CircularProgressIndicator(),
+    } else {
+      return FutureBuilder<bool>(
+        future: _init(),
+        initialData: false,
+        builder: (context, snapshot) {
+          if (snapshot.hasData && snapshot.data) {
+            return widget.child;
+          }
+          return widget.appLoader ??
+              MaterialApp(
+                theme: ThemeData(
+                  primarySwatch: Colors.blue,
+                ),
+                debugShowCheckedModeBanner: false,
+                home: Scaffold(
+                  backgroundColor: Colors.white,
+                  body: Center(
+                    child: SizedBox(
+                      height: 36,
+                      width: 36,
+                      child: CircularProgressIndicator(),
+                    ),
                   ),
                 ),
-              ),
-            );
-      },
-    );
+              );
+        },
+      );
+    }
   }
 }
 
@@ -1175,6 +1252,26 @@ class Momentum extends InheritedWidget {
     return null;
   }
 
+  String _validateInjectService(List<MomentumService> services) {
+    var injectedServices = services.where(
+      (s) {
+        return trycatch(() => (s as InjectService)._alias) != null;
+      },
+    ).cast<InjectService>();
+    for (var inject in injectedServices) {
+      var count = injectedServices
+          .where(
+            (x) => x._alias == inject._alias,
+          )
+          .length;
+      if (count > 1) {
+        return '[$Momentum] -> Duplicate alias is found. You passed '
+            'in $count "${inject._alias}"';
+      }
+    }
+    return null;
+  }
+
   final List<MomentumController> _controllers;
 
   final List<MomentumService> _services;
@@ -1221,17 +1318,54 @@ class Momentum extends InheritedWidget {
   }
 
   /// Method for testing only.
-  T serviceForTest<T extends MomentumService>() {
-    return _getService<T>();
+  T serviceForTest<T extends MomentumService>({dynamic alias}) {
+    return _getService<T>(alias: alias);
   }
 
-  T _getService<T extends MomentumService>() {
+  T _getService<T extends MomentumService>({dynamic alias}) {
     var type = _getType<T>();
-    var service = _services.firstWhere(
-      (c) => c.runtimeType == type,
-      orElse: () => null,
-    );
-    if (service == null) {
+    if (type == _getType<InjectService<MomentumService>>()) {
+      throw _invalidService;
+    }
+    if (type == _getType<InjectService<dynamic>>()) {
+      throw _invalidService;
+    }
+    T result;
+    if (alias == null) {
+      result = _services.firstWhere(
+        (c) => c.runtimeType == type,
+        orElse: () => null,
+      );
+      if (result == null) {
+        var injectors = _services
+            .where(
+              (s) => s.runtimeType == _getType<InjectService<T>>(),
+            )
+            .cast<InjectService>()
+            .toList();
+        if (injectors.isNotEmpty) {
+          result = injectors
+              .firstWhere(
+                (i) => i._service.runtimeType == type,
+              )
+              ?._service;
+        }
+      }
+    } else {
+      var injectors = _services
+          .where(
+            (s) => s.runtimeType == _getType<InjectService<T>>(),
+          )
+          .cast<InjectService>()
+          .toList();
+      result = injectors
+          .firstWhere(
+            (i) => i._alias == alias && i._service.runtimeType == type,
+            orElse: () => null,
+          )
+          ?._service;
+    }
+    if (result == null) {
       if (_testMode && type == _getType<InMemoryStorage>()) {
         return InMemoryStorage() as T;
       }
@@ -1239,7 +1373,7 @@ class Momentum extends InheritedWidget {
           'was not initialized from the "services" parameter '
           'in the Momentum constructor.');
     }
-    return service as T;
+    return result;
   }
 
   T _getControllerOfType<T extends MomentumController>([Type t]) {
@@ -1322,8 +1456,11 @@ class Momentum extends InheritedWidget {
   /// The static method for getting services inside a widget.
   /// The service must be marked with [MomentumService] and
   /// injected into [Momentum] root widget.
-  static T service<T extends MomentumService>(BuildContext context) {
-    return _getMomentumInstance(context)._getService<T>();
+  static T service<T extends MomentumService>(
+    BuildContext context, {
+    dynamic alias,
+  }) {
+    return _getMomentumInstance(context)._getService<T>(alias: alias);
   }
 
   static T _ofType<T extends MomentumController>(BuildContext context, Type t) {